Tabla ordenada Recorrer tabla

+++ src/cohete.java
@@ -1,6 +1,6 @@
 public class cohete extends Aeronave implements Propulsor{
 
-    public cohete(String codigo, String fechaentrega, String LG2A, int horastrabajoprev, int horastrabajoempleada, int capacidadpasajeros, double coste, double precioventa){
+    public cohete(String codigo, String fechaentrega, String LG2A, int horastrabajoprev, int horastrabajoempleada, int capacidadpasajeros, float coste, double precioventa){
         super(codigo,fechaentrega,LG2A,horastrabajoprev,horastrabajoempleada,capacidadpasajeros,coste,precioventa);
